Refactored Code to follow N-Tier Architecture
