+++ src/main/java/com/artchristian/customer/CustomerDao.java
@@ -1,0 +1,9 @@
+package com.artchristian.customer;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface CustomerDao {
+    List<Customer> selectAllCustomers();
+    Optional<Customer> selectCustomerById(Integer id);
+}
