+++ src/main/java/com/artchristian/customer/CustomerService.java
@@ -1,0 +1,27 @@
+package com.artchristian.customer;
+
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class CustomerService {
+
+    private final CustomerDao customerDao;
+
+    public CustomerService(CustomerDao customerDao) {
+        this.customerDao = customerDao;
+    }
+
+    public List<Customer> getAllCustomers(){
+        return customerDao.selectAllCustomers();
+    }
+
+    public Customer getCustomer(Integer id){
+        return customerDao.selectCustomerById(id)
+                .orElseThrow(
+                        ()-> new IllegalArgumentException(
+                                "Customer with ID [%s] is not found".formatted(id)));
+    }
+}
+
